+++ interfaces/food_handler.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 	"food-app/application"
 	"food-app/domain/entity"
-	"food-app/utils/auth"
-	"food-app/utils/fileupload"
+	"food-app/infrastructure/auth"
+	"food-app/interfaces/fileupload"
 	"github.com/gin-gonic/gin"
 	"net/http"
 	"os"
