+++ interfaces/food_handler_test.go
@@ -6,7 +6,7 @@ import (
 	"errors"
 	"fmt"
 	"food-app/domain/entity"
-	"food-app/utils/auth"
+	"food-app/infrastructure/auth"
 	"github.com/gin-gonic/gin"
 	"github.com/stretchr/testify/assert"
 	"io"
